refactor: canstants 파일 네이밍
