+++ src/advanced/constants/textContent.js
