+++ src/advanced/constants/timerConfig.js
